+++ src/main/java/com/nimble/server_spring/infra/security/CustomAuthenticationProcessingFilter.java
@@ -1,10 +1,12 @@
 package com.nimble.server_spring.infra.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nimble.server_spring.infra.http.RequestBodyParser;
 import com.nimble.server_spring.modules.auth.dto.request.LocalLoginRequestDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
@@ -28,8 +30,16 @@ public class CustomAuthenticationProcessingFilter extends AbstractAuthentication
         HttpServletRequest request,
         HttpServletResponse response
     ) throws AuthenticationException {
-        LocalLoginRequestDto localLoginRequestDto = RequestBodyParser.from(request)
-            .parseTo(LocalLoginRequestDto.class, objectMapper);
+        LocalLoginRequestDto localLoginRequestDto;
+        try {
+            localLoginRequestDto = RequestBodyParser.from(request)
+                .parseTo(LocalLoginRequestDto.class, objectMapper);
+        } catch (Exception exception) {
+            throw new BadCredentialsException(
+                "request body를 파싱하는데 실패했습니다. - " + exception.getMessage(),
+                exception
+            );
+        }
 
         return getAuthenticationManager().authenticate(
             new UsernamePasswordAuthenticationToken(
